changes made

+++ deletepost.php
@@ -6,41 +6,74 @@
 		$response=array();
 
 		$post_id = $_POST['post_id'];
+		$user_id = $_POST['user_id'];
 
-		$sql1 = "SELECT * FROM alumnus WHERE email=$_SESSION['email']";
-		$result1 = query($sql);
+		$sql = "SELECT * FROM posts WHERE id=$post_id";
+		$result = query($sql);
+		$row1 = fetch_array($result);
+		$user_id_post = $row1['user_id'];
+
+		$sql1 = "SELECT * FROM alumnus WHERE user_id=$user_id_post";
+		$result1 = query($sql1);
 		$row = fetch_array($result1);
 		$access_token = $row['access_token'];
 
-		$sql = "DELETE FROM posts WHERE id=$post_id";
-		$no=false;
 
-		$result = query($sql);
-        if($access_token==$_SESSION['access_token']){
-		 if(row_count($result)==1){
-		 $no=true;
-		 $sql1 = "DELETE FROM threads WHERE post_id=$post_id";
-		 $result1 = query($sql1);
-        
-	     }
-
-	     if($no){
-	     	$response['status']=209;
-	   $messages[]="post deleted succsesfully!";
-			
+		if($user_id==$user_id_post && $access_token==$_SESSION['access_token'])
+		{
+			$sql = "DELETE FROM posts WHERE id=$post_id";
+			$result = query($sql);
+			if(row_count($result)==1)
+			{
+				$sql1 = "DELETE FROM threads WHERE post_id=$post_id";
+				$result1 = query($sql1);
+				$response['status']=209;
+				$messages[]="post deleted succsesfully";
+			}else{
+				$response['status']=408;
+				$messages[]="something Went wrong retry!!";
+			}
 
 		}
-		 if($no){
-		 	$response['status']=408;
-			$messages[]="retry";
+		else
+		{
+			$response['status']=409;
+			$message[]="unauthorized access";
+
+		}
+
+
+
+
+	// 	$sql = "DELETE FROM posts WHERE id=$post_id";
+		
+
+	// 	$result = query($sql);
+ //        if($access_token==$_SESSION['access_token']){
+	// 	 if(row_count($result)==1){
+	// 	 $no=true;
+	// 	 $sql1 = "DELETE FROM threads WHERE post_id=$post_id";
+	// 	 $result1 = query($sql1);
+        
+	//      }
+
+	//      if($no){
+	//      	$response['status']=209;
+	//    $messages[]="post deleted succsesfully!";
+			
+
+	// 	}
+	// 	 if($no){
+	// 	 	$response['status']=408;
+	// 		$messages[]="retry";
 			
 
 	
-		}
-	}else{
-		$response['status']=409;
-		$message[]="unauthorized access";
-	}
+	// 	}
+	// }else{
+	// 	$response['status']=409;
+	// 	$message[]="unauthorized access";
+	// }
         
     
 	$response['messages'] = $messages;
